change stack-os-matrix to invoke sudo with a login shell
It seems that git is unable to clone a new repo if it does not have read
permission to the PWD (unrelated to the path of the clone).  This can be
avoided by having sudo invoke lsstswBuild.sh under a login shell so that the
PWD is changed to the homedir of the build role account.

git error message:

    fatal: Could not change back to '/home/jenkins-slave/workspace/stack-os-matrix/label/centos-7': Permission denied