add lsst_ci to stack-os-matrix job
By default, `lsst_ci` will be appended to the list of EUPS products to
build.  If the `SKIP_DEMO` option set to `true`, the product list is not
modified.  The intent is to treat `lsst_ci` similar to
`lsst_dm_stack_demo`, with the notable exception(s) of the later always
being run from the `master` branch and having its output present in the
console log.