set all request == limit; change swarm to 8 cores
This is to prevent load spikes that are reportedly disturbing other loads.
The swarm memory request/limit has been set to 64GiB to allow for large number
of build processes to be spawned based on the number of visisble cores, even
though the cgroup control should limit usage to 8 cores.